# server/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Mean Application API",
    "description": "Mean Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:4949",
  "basePath": "/api/",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    },
    {
      "name": "Auth",
      "description": "API for auth in the system"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "AuthHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Login to the system",
        "parameters": [
          {
            "name": "auth",
            "in": "body",
            "description": "User auth details",
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User is loggedin",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": [
          "Product"
        ],
        "responses": {
          "200": {
             "schema":{
              "$ref":"#/definitions/Products"
             }
          }
        }
      }
    },
    "/products/create": {
      "post": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "Create Product Detales",
            "schema": {
              "type": "object",
              "required": [
                "price",
                "balance",
                "name",
                "barcode"
              ],
              "properties": {
                "balance": {
                  "type": "number"
                },
                "price": {
                  "type": "number"
                },
                "name": {
                  "type": "string"
                },
                "barcode": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "description": "products"
      }
    },
    "/products/{id}": {
      "put": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "Create Product Detales",
            "schema": {
              "type": "object",
              "required": [
                "_id",
                "price",
                "balance",
                "name",
                "barcode"
              ],
              "properties": {
                "_id": {
                  "type": "string"
                },
                "balance": {
                  "type": "number"
                },
                "price": {
                  "type": "number"
                },
                "name": {
                  "type": "string"
                },
                "barcode": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "description": "products"
      }
    }
  },
  "definitions": {
    "User": {
      "required": [
        "email",
        "fullname"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItems": true
        },
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "fullname": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "roles": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Users": {
      "type": "array",
      "$ref": "#/definitions/User"
    },
    "Auth": {
      "type": "object",
      "properties": [
        {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/definitions/User"
          }
        }
      ]
    },
    "Product": {
      "type": "object",
      "properties": [
        {
          "_id":{
            "type":"string"
          },
          "name": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "price":{
            "type": "number"
          },
          "balance":{
            "type": "number"
          }
        }
      ]
    },
    "Products": {
      "type": "array",
      "properties": [
        {
          "$ref":"#/definitions/Product"
        }
      ]
    }
  }
}